refactor: extract SCM functions from iterationEngine into iterationScm
Moves 362 lines of git/branch-per-task SCM logic (commit-on-done,
merge, checkout, conflict detection, PR creation) into a focused
iterationScm.ts module. iterationEngine.ts reduced from 2074→1711 lines.
